pyproject.toml
[project]
name = "backend-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiomysql>=0.2.0",
    "celery>=5.4.0",
    "fastapi[standard]>=0.116.1",
    "pyjwt>=2.10.1",
    "redis>=6.2.0",
    "ruff>=0.12.7",
    "python-dotenv>=1.0.0",
    "itsdangerous>=2.2.0",
    "pydantic-settings>=2.10.1",
    "sqlmodel>=0.0.24",
    "pymysql>=1.1.0", # 同步驱动
    "greenlet>=3.0.0",
    "alembic>=1.16.4",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "argon2-cffi>=25.1.0",
    "authlib>=1.6.1",
    "dashscope>=1.24.1",
    "python-slugify>=8.0.0",
    "boto3>=1.40.11",
    "stripe>=12.4.0",
    "weasyprint>=66.0",
    "pycryptodome>=3.23.0",
    "qrcode[pil]>=8.2",
    "azure-cognitiveservices-speech>=1.46.0",
    "loguru>=0.7.3",
]

[tool.ruff]

exclude = [
    "migrations",
    "tests",
    "app/models/__init__.py",  # 忽略__init__.py文件中的某些规则
]

fix = true
